Add tests and update queries for various entities in the hospital management system - Updated Makefile to include a test target for running tests. - Modified database migration to change `id` type from INT to BIGINT in EmergencyLine table. - Added new test files for CheckUpTime, Doctor, EmergencyLine, OPDLine, and DoctorSpeciality with comprehensive test cases. - Implemented counting and searching functionalities - Added utility function to create sql.NullInt32 for age handling.

+++ db/migration/000001_init_schema.up.sql
@@ -74,7 +74,7 @@ CREATE TABLE `OPDLine` (
 );
 
 CREATE TABLE `EmergencyLine` (
-  `id` INT AUTO_INCREMENT PRIMARY KEY,
+  `id` BIGINT AUTO_INCREMENT PRIMARY KEY,
   `reg_time` DATETIME,
   `token_number` BIGINT NOT NULL,
   `client_id` BIGINT NOT NULL,
